worker/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
aio-pika==9.5.4
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-instrumentation-aio-pika==0.50b0
opentelemetry-instrumentation-logging==0.50b0
